requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-WTF==1.2.1
python-dotenv==1.0.1
PyMySQL==1.1.0
Werkzeug==3.0.1
email-validator==2.1.0.post1
cryptography==42.0.2
Pillow==10.2.0
python-dateutil==2.8.2
Flask-Cors==4.0.0
bcrypt==4.1.2
requests==2.31.0
pymongo==4.6.1
pymysql==1.1.0
numpy==1.26.3
plotly==5.18.0
gunicorn==21.2.0 